prjfs: correctly handle directory rename
Summary:
When a directory is renamed, EdenFS receives a single notification on that
directory. Unfortunately, due to the implementation of the fileNotification
function, this means the old directory would get removed, and the destination
directory would just get created. However, none of the files or other
directories would get added to the inode hierarchy.

To fix this, we need to make sure to recursively add the content of a directory
when it gets created as some files may be present in it.

Reviewed By: fanzeyi

Differential Revision: D34731270

fbshipit-source-id: 39b25441408c1e44ca7e62522699b053dd3e6c5d