Fix file/dir clashes cleanup and add tests
Summary:
ASIC reported errors on snapshots, which happen when the snapshot is trying to cleanup the working directory, but fails due to file/dir clashes. This is a problem more generic to hg and I've posted about it.

However, this diff adds a snapshot specific fix, by doing the cleanup in a very specific order to avoid errors. Ideally this would be fixed in the "revert" function and we won't need to be so careful in the future.

Reviewed By: markbt

Differential Revision: D34820751

fbshipit-source-id: 3ec13b4932fadd39cc89cf75475434226f2f781b