init: return repo object from initialization
Summary:
We are currently in the process of moving the clone command (and other parts of our codebase) into Rust. As part of this rewrite process, we will make some of the `localrepo` object delegate some tasks to the Rust `repo` object, and in order to do this we need some reference to it.

This change makes the Rust repo initialization return a `repo` object, modifies the bindings for it, and makes the `localrepo` hold a Rust `repo` object.

Reviewed By: DurhamG

Differential Revision: D34773297

fbshipit-source-id: 22f38fb0178fe7dce300077b515e95eb007d94bb