redaction: ensure admin tools use correct form of redaction
Summary:
Redaction's SQL implementation has been replaced by one that uses the
blobstore. Ensure the admin commands also use the new implementation when
querying the blobstore directly.

The old admin tool currently duplicates initialization (which is why this was missed
in the first place), but that will be removed when we remove the tool.

Reviewed By: yancouto

Differential Revision: D34790120

fbshipit-source-id: 45573864be86bda9669943a07dc9fbf84bf88beb