lfs: allow FnMut write_to_store functions
Summary:
In a later diff we'll want to call 'self.found_attribute' when we
receive a LFS file, to avoid having to pend things in memory. This requires
capturing '&mut self'. To make this possible, let's refactor the LFS fetcher to
be FnMut not require Send+Clone+'static by turning the futures into a stream and
offloading it to a background thread with stream_to_iter, then iterating over
that and calling write_to_store on the main thread.

Note, this does mean we won't be calling write_to_store from multiple background
threads in parallel anymore. I don't believe this matters though, since under the hood
the indexedlog only allows one writer at a time anyway.

Reviewed By: yancouto

Differential Revision: D34803638

fbshipit-source-id: 9437a529136787a103e70f58eed335c956f93c8f