cli: allow eden rm to remove unknown directories that failed cleanup
Summary:
Some times, when user failed their `eden rm` on Windows because of files being in use by other process. They would try to run the same `eden rm` command again. It wouldn't work because the mount point is no longer an EdenFS directory so we bail out.

This has confused our user a bit and this diff will allow `eden rm` to detect if the path is a previously EdenFS mount on Windows, and continue the clean up process.

Reviewed By: xavierd

Differential Revision: D34794411

fbshipit-source-id: eaa86390257412daf80908bf88559b8f639c7a36