cli: make handle.exe check interruptible
Summary:
`handle.exe` would take several seconds to finish (~20s on my machine) and sometimes the user may not want to wait for it to finish. This will allow the user to safely interrupt its invocation and we still print the command to run if they want to find out which processes are using files from the repo themselves.

Also some small cosmetic changes.

Reviewed By: xavierd

Differential Revision: D34802959

fbshipit-source-id: 4d9b53b7feea1e4378b46c325d2e0adfd98aca74