status: implement TreeState-based status logic
Summary:
This diff adds the initial Rust logic for the status command. It combines the data from the pending filesystem changes with the current TreeState to determine what in the filesystem has changed relative to the actual tree.

The full status logic handles more cases (i.e. things that don't actually depend on the pending filesystem changes). Not all of these are necessarily applicable to a Rust rewrite of the status logic, but the other needed cases will be added later.

Reviewed By: DurhamG

Differential Revision: D34768969

fbshipit-source-id: d64532c7b8a6472b243558b7aeacf7e8340de1bf