status: Handle files not in "pending changes"
Summary:
The previous diff does the majority of the status work: looking at each file in "pending changes" and using that to determine the status of each file based on the TreeState.

However, there's additional logic that needs to be done. Pending changes only shows differences between P1 and the working copy. That means that files are aren't in P1 and aren't in the working copy won't be in pending changes. This matters in a few cases:
1) files added to the tree and then deleted on the filesystem (should be "deleted")
2) files from P2 (not in P1) that are `hg remove`'d (should be "removed")
3) files from P2 (not in P1) that are deleted on the filesystem (should be "deleted").
4) clean files that have retroactively been marked as copied (i.e. with `hg cp --force --after`)

We have an additional constraint: at this layer, we can't do any more filesystem operations, since we should be independent of the underlying filesystem. Thus, we have to be clever in terms of how we combine the pending changes information with the tree state.

The exact logic (especially for the P2 cases, which is a bit more complex) is explained in a comment in the code, but we can use the fact that files that aren't in pending changes and aren't in P1 must not exist on the filesystem to handle these cases.

Reviewed By: DurhamG

Differential Revision: D34806635

fbshipit-source-id: 4cc35810e8954e22f373bb2e99482ae4eefdf67a