status: remove 'listunknown' parameter
Summary: This diff removes the 'listunknown' parameter from the Rust status code. Computing the unknown files (once we already have "pending changes") is linear, and generally something we want to be computing. This simplifies the API, and keeps the old behavior in the dirstate code that calls into Rust by emptying the list if we don't want it.

Reviewed By: DurhamG

Differential Revision: D34832721

fbshipit-source-id: f5b8e791937af81e8766b6fa03bf5733569c2444