checkout: remove conflicting ancestor symlinks
Summary:
Now "checkout --force" will remove unknown ancestor symlinks that conflict with a file write. For example, you can now write out a file "a/b" if "a" is currently a "?" symlink.

We already had code to delete symlinks to prepare for a write in workingfilectx.clearunknown and vfs::Vfs::clear_conflicts, but both first audited the path which errored out because the path contained a symlink. I tweaked clear_conflicts to still perform symlink cleanup even if the path doesn't audit. I tried to be conservative and gate the other work behind a succesful audit.

A user hit this bug when an interrupted checkout left them with a "?" symlink that the checkout destination wanted to replace with a non-symlink. Even "up -C" could not clean it up.

Reviewed By: DurhamG

Differential Revision: D34542766

fbshipit-source-id: b55f6c117bcd762e4683bb9c0f1681310c226e77