fuse: early return from flushInvalidation when channel is unmounted
Summary:
When unmount and checkout are both in progress at the same time, the
FuseChannel may be unmounted prior to the checkout operation completing in that
case, the invalidation thread would be terminated prior to checkout completing.
This would also mean that checkout wouldn't be able to complete due to waiting
on the invalidation thread to flush all the invalidations, but since that
thread is gone, the flush would never complete.

This doesn't affect NFS due to the NFS code relying on a folly Executor and not
a separate thread.

Reviewed By: chadaustin

Differential Revision: D34905149

fbshipit-source-id: e0358f28e6a38dd2b0a14d9a6b5071389ad68b62