networkdoctor: start diagnosing http level problems
Summary:
Now the network doctor also attempts to make http requests to suss out what is wrong. We request /edenapi/:repo/capabilities since this will trigger permission errors (as opposed to /health_check). It tries with and without x2pagentd (if configured) to further triangulate the source of the problem.

This commit only contains high level categorization of the problem. I will follow up with the "diagnosis" which will start to map HTTP responses into informational messages for users.

Reviewed By: DurhamG

Differential Revision: D34734186

fbshipit-source-id: 794c5dd8a89d6d0cfc2068b8e3f5cc98e6b7959b