remove taggederror crate and related code
Summary:
taggederror wasn't used and doesn't seem like the direction we want to go in, so excise it.

taggederror allowed for error tagging and error metadata, and was wired around in various places. It tries to solve real problems, but it was too quick to prescribe how things should work without applying the design to see if it works in practice.

Reviewed By: DurhamG

Differential Revision: D34649753

fbshipit-source-id: bc60896f4e1dbe3feb7d5804561392d1f1ebda72