pathmatcher: more efficient tree-based ExactMatcher
Summary: This diff improves the core functionality of the ExactMatcher in the previous diff to intelligently return whether or not a directory needs to be traversed by checking an internal tree structure to see if any of the matcher's files are in that directory. This should improve the performance of Matcher-based tree traversal.

Reviewed By: DurhamG

Differential Revision: D34912969

fbshipit-source-id: 51349938b3a58de5462a77f0519c021014f3b228