tests/utils: migrate to facet-based repos
Summary:
To allow tests that use the `tests_utils` crate to work with repos other than
`BlobRepo`, migrate this crate to using facet reference definitions.

In a lot of tests, this was the only thing that required `BlobRepo`, which
means type inferencing now does not know which concrete type to use.  In these
cases, we specify `BlobRepo` for now.  In the future these can be changed to a
simpler repo type for that test.

Reviewed By: yancouto

Differential Revision: D34864664

fbshipit-source-id: e4142e8e47c3e942a5d9ceac4ba4270fe597f097