prjfs: remove the dependency on RCU
Summary:
Our use of RCU is fundamentally incompatible with N4895 since it allows a read
section to be entered in one thread and closed in another one. We use this
feature in Prjfs to prevent a mount to be unmounted while requests are ongoing
on it.

Let's instead make use of a folly::AtomicReadMostlyMainPtr which is equivalent
to a folly::Synchronized<std::shared_ptr<>> without requiring as many atomics
in the read path.

Reviewed By: Decave

Differential Revision: D34982722

fbshipit-source-id: 299f25fc8c8aa72fb89cc9fc880860780ea0517d