mercurial_derived_data: make get_hg_changeset_from_bonsai an extension trait
Summary:
Make the free function in `mercurial_derived_data` into an extension trait that will work
on any kind of repo that supports derived data.

The ordinary way to get a mercurial changeset with facet-based repos would be:
```
   use repo_derived_data::RepoDerivedDataRef;
   use mercurial_derived_data::MappedHgChangesetId;

   let hg_changeset_id = repo
        .repo_derived_data()
        .derive::<MappedHgChangsetId>(ctx, cs_id)
        .await?
        .hg_changeset_id();
```

However, this is a bit of a mouthful for something that is very common in Mononoke.  Instead,
make it similar to how `BlobRepo` works by turning it into an extension trait:
```
   use mercurial_derived_data::DeriveHgChangeset;

   let hg_changeset_id = repo.derive_hg_changeset(ctx, cs_id).await?;
```

The method name is changed to make it clear that derivation may be triggered,
and `BlobRepo::get_hg_from_bonsai_changeset` is changed to use this new
method.

Reviewed By: mitrandir77

Differential Revision: D34971450

fbshipit-source-id: 890d106278e339c2a6deb973eea8ed5122baa4f8