prjfs: reduce the time directories are kept opened
Summary:
Thanks to ImmediateFuture not deferring anything in the fast path,
recursivelyAddAllChildrens may recursively add all the childrens without
sleeping or returning control to the executor. The downside is that directories
can be held open for a long time, and on Windows, this would prevent subsequent
operations on the directories themself.

Unity in particular is found of creating directories, adding files, renaming
the directory, etc in a very fast fashion. And in some cases, EdenFS would hold
the directory opened for too long, raising an error in Unity which would affect
users. To reduce this, we can simply collect all the directory entry names
first and then recursively handle them after the directory is closed.

Reviewed By: chadaustin

Differential Revision: D35018455

fbshipit-source-id: 8f074ff794c64d2a65607fc0829ec444b80ddb9a