runlog: infer command boringness in some cases
Summary:
Now the runlog infers than an hg command is "boring" if the command disables the blackbox. "boring" means the runlog won't touch the watchfile to notify VSCode, and runlog will more aggressively clean up the runlog files.

This is not a good solution, but disabling the blackbox seems to be the only indicator that commands are known spammy. We should probably introduce an explicit option like that can inform the blackbox, runlog and other things that want to separate interesting commands from spam commands.

Reviewed By: quark-zju

Differential Revision: D34979231

fbshipit-source-id: 5aa4291360b71fdab5efa97514977acb085c80e0