mononoke: refactor test fixtures library to convert fixtures from modules to structs
Summary:
Having each repo fixture as a separate module with repetitive code looks wrong. The repetitive code is also inconsistent between different fixtures.

I introduced a proper trait with default function implementations and replaced each module with a struct. The end code is shorter and cleaner, but I needed to modify a lot of other files because of naming changes and `use fixtures::TestRepoFixture`.

Reviewed By: markbt

Differential Revision: D35081002

fbshipit-source-id: fa753d933665afcd89cd902b20205de1968de29c