remotefilelog: simplify repo.prefetch
Summary:
The old code had support for patterns, matcher options, and a matcher.
Most of that was never used, and all of it can be expressed via a matcher, so
let's simplify the interface to just contain a matcher.

Also, the old code had a notion of force fetching from the server. This was an
artifact of the old way history worked, where we may need to overwrite local
history with remote history. This is no longer required, since we now adjust
link revs. Let's remove it too.

Reviewed By: quark-zju

Differential Revision: D35033073

fbshipit-source-id: b204ebd30cb1bae7caa0030d6672f415caeb4353