status: make EdenFS repos work with Rust status
Summary:
This diff allows EdenFS-backed repos to use the `workingcopy.ruststatus` parameter. The native Rust status implementation only supports the new `TreeState` for the state of the working copy. This is fine for many repos, however, `EdenFS`-backed repos still use the old `dirstate` format.

This diff adds a bit of a hack to convert EdenFS dirstates temporarily into a read-only copy in a TreeState in a temporary directory, and then pass that to the Rust status. The previous diff in this stack already adjusted the Rust status code to support `TreeState`s that only list non-normal files (e.g. added and removed).

At some point if/when EdenFS switches to use the new TreeState format to store this information, this code can be changed.

Reviewed By: DurhamG

Differential Revision: D35098347

fbshipit-source-id: cd6fbcc81e7026fbb00963ffff6949d6b4b00786