status: lookup missing files in the TreeManifest
Summary:
Previously, we used the TreeState as the source of truth for all information about files in the tree. This works fine for physical repos. However, EdenFS repos (not yet supported by the Rust status code) are lazy, and only hold information about added/removed paths in their TreeState (rather than containing all tracked files, which may be quite large). The first version of this code made the assumption that any file not in the TreeState was necessarily untracked.

This diff updates the code to, if a file isn't in the TreeState, add it to a list to be looked up in the TreeManifest, which is the *actual* source of truth about whether or not a file is tracked in the repo. This adds a bit of complexity, because we want to query the TreeManifest efficiently (e.g. in a large batch, with BFS, pruned with an ExactMatcher).

Reviewed By: DurhamG

Differential Revision: D34980676

fbshipit-source-id: 91263f48723255a85760bdfd1375bc10ad062054