pull: implement lazy pull for clone
Summary:
Add crate for pull functionality and implement a method for pulling commit and bookmarks data for clone using the bookmarks + lazy_pull endpoints. This requires a lazy changelog repo.

Notes: this only writes remotenames & tip to the metalog, but not to its double file in the store. If we still want to support the double writing, we'll need to add that functionality to the metalog. Additionally, this only supports pulling selectivepull bookmarks that are on the master line.

Reviewed By: DurhamG

Differential Revision: D34741868

fbshipit-source-id: 81c6b5f91c55a46b894410b992fe50a18eaaf7ba