mononoke_app: rename mysql_max_query_time back to mysql_query_time_limit
Summary:
In D33622581 (https://github.com/facebookexperimental/eden/commit/0a771a579e79d08d19c9efd9571562e4385bc775) we added the new version of `MysqlArgs` to replace the old one
that used an older version of clap.

However, the `mysql_max_query_time` field should actually have been called
`mysql_query_time_limit` (the constant name and value differed on the old code,
which is what caused the mistake).

Rename the field to what it should've been.  This will allow callers that are
setting that option to continue to work with the new version of clap.  Keep
the other name as an alias as some callers have already been modified.

Reviewed By: yancouto

Differential Revision: D35079094

fbshipit-source-id: 1b5cfcb96216c47421b65028f8fe09d365c0172d