test: fix test-infinitepush-hydrated flakiness
Summary:
This test was using 'hg checkout X' to determine if the files in X were
already fetched before. But 'hg checkout X' first must do 'hg status', which may
need the contents of the current commit Y. There is a non-deterministic
component to hg status (based on file timestamp alignment), which I think is
causing the flakiness).

An easy fix is to stop using hg checkout, and instead use hg cat to specifically
test the file we expect to be on disk already.

Reviewed By: mzr

Differential Revision: D35161684

fbshipit-source-id: 06b0490d02ad02cfd55800a6842ca9fce350c02b