Move DM ops tests into test_utils
Summary:
**Context**: Setting up for implementing deleted manifest v2, but for now it's just refactoring to make DMv1 more generic and be able to reuse this later in v2.

This moves the tests in `ops.rs` to the `test_utils.rs` file. It does not make them generic yet, to make it easier to review since it's clear they're basically the same.

There was some duplicated code between the tests in the two files, so I was now able to reuse this and deleted some methods :)

Reviewed By: markbt

Differential Revision: D34759062

fbshipit-source-id: dd351366e603365e95ce28e09207fb407a19d014