Make DM tests generics over any DM
Summary:
**Context**: Setting up for implementing deleted manifest v2, but for now it's just refactoring to make DMv1 more generic and be able to reuse this later in v2.

This diff makes the tests in `test_utils` generic over the "deleted manifest common" traits. This will allow them to be reused by other DM implementations in the future (not this diff). It involved changing the tests to use the manifest through the `DeletedManifestCommon` trait, which luckily needed very few changes.

To do that, I had to add a `RootDeletedManifestIdCommon` trait to be implemented by "root ids" (the thing that's actually derived) for all manifest (and implement it only for v1 now)

I also moved the tests from the `derive` file to the `mapping` file because I thought it made more sense, as it created the `RootDeletedManifestId` type.

Reviewed By: markbt

Differential Revision: D34756725

fbshipit-source-id: 39325606992bff522efac72d419127aa39a00091