dump edenapi to scuba and then Hive
Summary:
This records sampled edenapi requests and records then to Scuba so they can be later replayed. There's automation that creates daily partitions in Hive based on data in Scuba and we actually indent to use Hive. This is the simplest way to log to Hive from Rust now. Propper Logger support will come by the end of H1, then I will change it to use Logger here.

I will need to modify tunables in configerator to actually turn this on.

Reviewed By: mitrandir77

Differential Revision: D34418355

fbshipit-source-id: 8ed185b100c90e4f65f4bd7d10055d28ac094a62