snapshots: best effort snapshot on conflict
Summary: Right now we fail to create a snapshot when in the middle of a rebase conflict, as we don't support conflicts or merges. This diff makes it so we create a snapshot that has all the correct file content, but not the actual conflict information, as that needs significant changes.

Reviewed By: quark-zju

Differential Revision: D35182690

fbshipit-source-id: 8af9a528cd84e8e29a1abf85692030a45f17956d