add version and uptime info to E-Menu
Summary:
This adds an "About Eden" section to the E-Menu which lists the current running EdenFS version and uptime.

We approximate the uptime by passing chrono::now to the Windows Notifier when we create it (it should be very-very-very close since we create the Notifier when we start the Daemon. Therefore, the uptime of the two should be equal).

Reviewed By: chadaustin

Differential Revision: D34968215

fbshipit-source-id: 726d1229b880e2ae10c2950ac4a50f36badfe9db