edenfs-client: change status fast-path to use RepoPath
Summary:
This diff changes the internals of the EdenFS status fast-path to use the `types::RepoPath` type internally for representing paths within the repository, instead of the `std::path::Path` type.

The motivation for this is that it will allow us to more easily interoperate with the Rust `status::Status` (supporting a future refactor where all of the printing logic is done in `hgcommands`). The only feature we're theoretically losing from this is invalid-UTF8 support (RepoPath must always be UTF-8, Path doesn't necessarily need to be). However, the status fast-path logic didn't support this to begin with, so we're not actually losing anything.

Reviewed By: quark-zju

Differential Revision: D35201663

fbshipit-source-id: 741b8d0348b59ce04f8368f0d7aa3522717c6f46