hgcommands: move Rust status command printing to hgcommands from edenfs-client
Summary:
Previously, the EdenFS status fast-path logic for printing the status (e.g. with color, with or without copy information, symbols, etc) was all implemented in the `edenfs-client` crate.

This diff moves the printing logic (and the tests) over to the `hgcommands` crate, so that in the future, a fully native Rust implementation can easily re-use the printing logic.

Reviewed By: quark-zju

Differential Revision: D35227559

fbshipit-source-id: 47dbe2af0dbfcf70034e46b0fcaa4a2e531f73f3