edenfs-client: refactor status to use the Rust status crate
Summary:
This diff refactors the implementation of the EdenFS status fast path to use `status::Status` to keep track of status, and splits up the internal functionality in a way that makes it easier to move the printing logic to `hgcommands`.

The `Status` struct uses a HashMap internally and doesn't preserve ordering, so we `sort` the status output of each group before outputting for display purposes.

A future diff will split up the tests between "grouping" (the logic that turns raw EdenFS output to a Status), and "printing", in preparation for future changes.

Reviewed By: quark-zju

Differential Revision: D35202773

fbshipit-source-id: a56e94a40e101aed308b14105819f8766f2f7128