cli: create symlink redirections atomically on Windows
Summary:
Symlink creation on Windows is non-atomic, that is the symlink is first created
as a directory, and then a reparse tag is added to it. For EdenFS, this means
that when the notification thread queries the filesystem, it may either see
that the symlink is a directory, or a symlink, depending on whether EdenFS wins
the race.

If EdenFS finds out that this is a directory, it will try to recursively add
every file and directory under it to the inode hierarchy, which can be
extremely slow. Some users have noticed new clones taking many minutes to
answer a simple `hg status` query due to that query waiting on all pending
notifications to be completed and EdenFS is still processing the symlink.

To solve this in an ad-hoc way, we can force a symlink to be created atomically
by first creating it outside of the repository, and then moving it in the
repository. The rename operation is atomic, and thus EdenFS always detect it as
a symlink!

Reviewed By: jtbraun

Differential Revision: D35263809

fbshipit-source-id: e6794976551838bfaba994e83f610d6c776422a2