metalog: improve default conflict handling
Summary:
Metalog currently conservatively rejects any changes if this != ancestor:

  this (content buffered in memory to commit)
   |
   |   other (during commit, metalog changed on disk)
   |     /
  ancestor (before commit, metalog loaded from disk)

However, a typical 3-way merge (such as the one used in pushrebase) is happy
to accept "non-conflict" changes (ex. only one side is changed, or both sides
make a same change). Update the metalog default resolver to do that.

This is not an issue for hg since hg still uses repo locks to prevent racy
writes to metalog. It is surfaced as an UX issue for git support since reading
the commit graph can trigger a write to the metalog without repo lock.

Differential Revision: D34885109

fbshipit-source-id: 8f9fda43633944fea50badc500addfafd7d68e89