get the absolute latest version when looking up a phab diff
Summary:
There are three choices for "version" in Phabricator:

* `latest_phabricator_version`, which gets V0.1 if there's no published version, or the latest published version ignoring drafts (V6 when there's a draft V6.3).
* `default_associated_native_phabricator_version`, which gets the latest published version if the diff is published (same as `latest_phabricator_version`), or the most recent draft if it's unpublished (so V0.9 instead of V0.1 if you're iterating on an unpublished diff).
* `latest_associated_phabricator_version_regardless_of_viewer`, which simply gets the latest version no matter what (so V6.2 if there are drafts waiting to become V7).

We currently use `latest_phabricator_version`, which is consistent with `jf get`, but means that if you're iterating on an unpublished diff, you get a stale version.

Switch to `latest_associated_phabricator_version_regardless_of_viewer` so that a `hg co Dxxxx` gets you the absolute most recent version. I would expect that reviewers would use `jf get`, which uses `latest_phabricator_version`, and thus not be surprised by this change. If this change does surprise people, we can decide what to do then.

Reviewed By: quark-zju

Differential Revision: D35287431

fbshipit-source-id: 1360cf799b21d90b3cde642d62886f0c4b4cd086