networkdoctor: flesh out top level messaging
Summary: Improve the message the users sees when the network doctor is run automatically due to a fatal network error. We now clearly state the command failed due to a network error, and we write the detailed output to a file. Previously we just showed the network doctor output and it wasn't clear the command had failed.

Reviewed By: quark-zju

Differential Revision: D35255833

fbshipit-source-id: 764abce8dcb7796283abc4b9fe1a485e04ff74c7