mononoke/cmdlib: implement walker common setup
Summary:
This is a refactoring/reimplementation of the [`setup_common(..)`](https://www.internalfb.com/code/fbsource/[e22d12de8703]/fbcode/eden/mononoke/walker/src/commands/setup.rs?lines=1308) function and which is heavily based on the old clap and cmdlib infrastructure. The function is huge and hard to read, I tried to split new one into the smaller functional pieces.

For some reason in the old setup the repos are initialized sequentially, I preserved behaviour in the new one, but intend to make it parallel in a separate diff.

Reviewed By: mitrandir77

Differential Revision: D35247782

fbshipit-source-id: aa5cd4ee2f221f62f988765445ae29d8c7289b45