Move BlobstoreKeyWrapper to typed_hash and implement for all hashes
Summary:
This is a different approach to D34790637.

The issue is that we can't implement `MononokeId` for some ids to come (ShardedMapNodeId), but we want to call the `blobstore_key` method still. When refactoring, I noticed `megarepo_api` already has a neat trait to do this: `BlobstoreKeyWrapper`.

So this diff implements `BlobstoreKeyWrapper` for all ids, and makes MononokeId depend on that instead and not implement `blobstore_key` manually.

Hopefully this approach is less complicated.

Reviewed By: markbt

Differential Revision: D35217399

fbshipit-source-id: 740bbdfaaf68b7e72c1a13987520ff2404d81035