Basic load/store in thrift + size/empty
Summary:
**Context**: Sharded maps are going to be used in DMv2 implementation.

This diff adds the basic boilerplate code for working with sharded maps in Rust. Algorithms are coming later.

It contains:
- Logic for converting to/from thrift for the Rust sharded map counterpart.
- `is_empty` and `size` methods, which are trivial.

One thing to notice: In Rust, the type is `ShardedMapNode<Value>`, where Value is any type that can be converted to/from binary. In thrift, there's no generics, so we store all values as binary. This is why D35217399 was needed, since we needed to decouple `blobstore_key` from `MononokeId`, as we can't implement that trait for `ShardedMapNodeId`.

Reviewed By: kris1319

Differential Revision: D35219036

fbshipit-source-id: ba4b137c4b60532a18b05ccc4385ff59362b50a8