doctor: add a Windows checker to check the sha1 of loaded files
Summary:
One of the issue some users have seen on Windows on a regular basis is when the
content of files on disk differs from what EdenFS thinks the content should be.
This leads to `hg status` not reporting the correct output and sometimes random
build failures. At first, let's try to detect these occurences. This diff focus
on only computing the sha1 of files that have been fully read already as
reading other files would force EdenFS to populate them on disk.

Reviewed By: kmancini

Differential Revision: D34776469

fbshipit-source-id: d04d31f7f7bfba85cbfbf68ce55f5490bbaec45f